refactor: elasticsearch clean up
